refactor(react-headless-components-preview)!: collapse Popover to a single auto-mode component
