+++ packages/react-components/react-headless-components-preview/library/etc/popover.api.md
@@ -29,13 +29,7 @@ export const Popover: {
 };
 
 // @public
-export const PopoverAuto: {
-    (props: PopoverProps): JSXElement;
-    displayName: string;
-};
-
-// @public
-export type PopoverContextValue = Pick<PopoverState, 'open' | 'setOpen' | 'toggleOpen' | 'triggerRef' | 'contentRef' | 'arrowRef' | 'openOnHover' | 'openOnContext' | 'disableAutoFocus' | 'withArrow' | 'inline' | 'mountNode' | 'popoverType'> & {
+export type PopoverContextValue = Pick<PopoverState, 'open' | 'setOpen' | 'toggleOpen' | 'triggerRef' | 'contentRef' | 'arrowRef' | 'openOnHover' | 'openOnContext' | 'disableAutoFocus' | 'withArrow' | 'inline' | 'mountNode'> & {
     positioning: {
         targetRef: React_2.RefCallback<HTMLElement>;
         containerRef: React_2.RefCallback<HTMLElement>;
@@ -54,8 +48,6 @@ export type PopoverProps = {
     positioning?: PositioningShorthand;
     withArrow?: boolean;
     disableAutoFocus?: boolean;
-    closeOnScroll?: boolean;
-    closeOnIframeFocus?: boolean;
     inline?: boolean;
     mountNode?: HTMLElement | null;
 };
@@ -78,7 +70,6 @@ export type PopoverState = Required<Pick<PopoverProps, 'open' | 'inline'>> & Pic
         y: number;
     } | undefined) => void;
     positioning: PositioningReturn;
-    popoverType: PopoverType;
 };
 
 // @public
@@ -116,9 +107,6 @@ export type PopoverTriggerState = {
 };
 
 // @public
-export type PopoverType = 'manual' | 'auto';
-
-// @public
 export const renderPopover: (state: PopoverState, contextValues: {
     popover: PopoverContextValue;
 }) => React_2.ReactElement;
@@ -131,9 +119,6 @@ export const renderPopoverTrigger: (state: PopoverTriggerState) => JSXElement | 
 
 // @public
 export const usePopover: (props: PopoverProps) => PopoverState;
-
-// @public (undocumented)
-export const usePopoverAuto: (props: PopoverProps) => PopoverState;
 
 // @public
 export const usePopoverContext: <T>(selector: ContextSelector<PopoverContextValue, T>) => T;
