+++ packages/react-components/react-headless-components-preview/library/src/components/Popover/PopoverSurface/usePopoverSurface.ts
@@ -8,10 +8,14 @@ import type { PopoverSurfaceProps, PopoverSurfaceState } from './PopoverSurface.
 
 /**
  * Returns the state for a PopoverSurface component.
+ *
+ * The surface is rendered with `popover="auto"` so the browser owns light
+ * dismiss (Escape, click-outside, popover-stack peer dismissal) — there is
+ * no React-side Escape handler. Hover handlers stay because hover is an
+ * open-side concern.
  */
 export const usePopoverSurface = (props: PopoverSurfaceProps, ref: React.Ref<HTMLDivElement>): PopoverSurfaceState => {
   const contentRef = usePopoverContext(context => context.contentRef);
-  const triggerRef = usePopoverContext(context => context.triggerRef);
   const openOnHover = usePopoverContext(context => context.openOnHover);
   const setOpen = usePopoverContext(context => context.setOpen);
   const arrowRef = usePopoverContext(context => context.arrowRef);
@@ -19,10 +23,7 @@ export const usePopoverSurface = (props: PopoverSurfaceProps, ref: React.Ref<HTM
   const inline = usePopoverContext(context => context.inline);
   const open = usePopoverContext(context => context.open);
   const mountNode = usePopoverContext(context => context.mountNode);
-  const popoverType = usePopoverContext(context => context.popoverType);
   const positioningCtx = usePopoverContext(context => context.positioning);
-
-  const browserHandlesDismiss = popoverType === 'auto' && !inline;
 
   const state: PopoverSurfaceState = {
     inline,
@@ -43,11 +44,7 @@ export const usePopoverSurface = (props: PopoverSurfaceProps, ref: React.Ref<HTM
     'data-open': open ? 'true' : 'false',
   };
 
-  const {
-    onMouseEnter: onMouseEnterOriginal,
-    onMouseLeave: onMouseLeaveOriginal,
-    onKeyDown: onKeyDownOriginal,
-  } = state.root;
+  const { onMouseEnter: onMouseEnterOriginal, onMouseLeave: onMouseLeaveOriginal } = state.root;
 
   state.root.onMouseEnter = useEventCallback((e: React.MouseEvent<HTMLDivElement>) => {
     if (openOnHover) {
@@ -61,20 +58,6 @@ export const usePopoverSurface = (props: PopoverSurfaceProps, ref: React.Ref<HTM
       setOpen(e, false);
     }
     onMouseLeaveOriginal?.(e);
-  });
-
-  state.root.onKeyDown = useEventCallback((e: React.KeyboardEvent<HTMLDivElement>) => {
-    if (!browserHandlesDismiss && e.key === 'Escape' && !e.isDefaultPrevented()) {
-      const target = e.target as HTMLElement;
-      const surface = contentRef.current;
-      if (surface && target.closest('[data-popover-surface]') === surface) {
-        e.preventDefault();
-        setOpen(e, false);
-        triggerRef.current?.focus();
-      }
-    }
-
-    onKeyDownOriginal?.(e);
   });
 
   return state;
